Move all sorted aggregation stuff to own files
