+++ cpp/arcticdb/processing/aggregation_utils.hpp
@@ -1,0 +1,16 @@
+/* Copyright 2023 Man Group Operations Limited
+ *
+ * Use of this software is governed by the Business Source License 1.1 included in the file licenses/BSL.txt.
+ *
+ * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
+ */
+
+#pragma once
+
+#include <arcticdb/entity/types.hpp>
+
+namespace arcticdb {
+
+void add_data_type_impl(entity::DataType data_type, std::optional<entity::DataType>& current_data_type);
+
+} // namespace arcticdb
